Cleanup and some improvements
Prevent power output from going to 100% when booting in Automated controller mode with the temperature mode set to External Input. The External Input will stay at -127.0 until updated via the API. If temperature is -127.0 then power output will be set to the minimum output setting.

Allow the AP mode SSID to be something other than CheapoDC. It will now match the hostname which is configurable in the WebUI.

Javascript cleanup of CDCjs.js

Documentation cleanup and typo fixes.